Move functions to modules Put results into nice format
